% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_birth_pop_states}
\alias{df_birth_pop_states}
\title{Birth data of Mexico, between 1970 to 2050}
\format{
A data frame with 2,673 rows and 5 variables:
\describe{
\item{year} {Year of the data}
\item{state} {State where the births occurred}
\item{CVE_GEO} {Geographical code of the state of occurrence}
\item{population} {Number of people living in that year and state}
\item{births} {Number of births that occurred in that year and state}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/proyecciones-de-la-poblacion-de-mexico-y-de-las-entidades-federativas-2016-2050/resource/284ef503-6c57-4a5c-8247-e13241ad4cee}
}
\usage{
df_birth_pop_states
}
\description{
A dataset containing the number of births estimated by CONAPO between 1970
and 2050. These estimations are disaggregated by year and state but not by
sex of the child.
}
\keyword{datasets}
