% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_births_INEGI}
\alias{df_births_INEGI}
\title{Birth data of Mexico, between 1985 to 2020, disaggregated by sex}
\format{
A data frame with 3,564 rows and 8 variables:
\describe{
\item{year} {Year of the data}
\item{state} {State where the births occurred}
\item{sex} {Sex of the child}
\item{CVE_GEO} {Geographical code of the state of occurrence}
\item{births} {Number of births that occurred in that year, state and sex}
\item{birth_prop} {Proportion of births in relation to the sum of both sexes}
\item{population} {Number of people living in that year and state}
\item{birth_rate} {Number of births divided by the population with the same year, state and sex.}
}
}
\source{
\url{https://www.inegi.org.mx/sistemas/olap/Proyectos/bd/continuas/natalidad/nacimientos.asp}
}
\usage{
df_births_INEGI
}
\description{
A dataset containing the number of births registered by INEGI between 1985
and 2020. These estimations are disaggregated by year and state and sex
}
\keyword{datasets}
