% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_migration}
\alias{df_migration}
\title{Migration data of Mexico, from 1970 to 2050 aggregated by ages.}
\format{
A data frame with 84,672 rows and 8 variables:
\describe{
\item{year} {Year of the data}
\item{state} {State where the migration occurred}
\item{CVE_GEO} {Geographical code of the state of occurrence}
\item{sex} {Sex of the migrants}
\item{age_group} {Age group to which belongs the data}
\item{type} {Type of migration: Interstate, International or Total}
\item{emigrants} {Number of emigrants}
\item{immigrants} {Number of immigrants}
\item{migratoin} {Net migration (immigrants - emigrants)}
\item{em_rate} {emigration rate}
\item{im_rate} {immigration rate}
}
}
\source{
\url{ https://datos.gob.mx/busca/dataset/proyecciones-de-la-poblacion-de-mexico-y-de-las-entidades-federativas-2016-2050/resource/81802225-00c4-49b4-ab49-a491334062b4,
 https://datos.gob.mx/busca/dataset/proyecciones-de-la-poblacion-de-mexico-y-de-las-entidades-federativas-2016-2050/resource/3f9c8392-153d-4299-a027-9f1b179f6edf}
}
\usage{
df_migration
}
\description{
A dataset containing the estimated migration by CONAPO between 1970 and 2050.
These estimations are disaggregated by year, state, sex and type of migration
(immigration or emigration). This dataset includes emigration rate estimates.
This dataset groups the information by groups of ages that increases by 5
years.
}
\keyword{datasets}
