% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_migration_expanded}
\alias{df_migration_expanded}
\title{Migration data of Mexico, from 1970 to 2050.}
\format{
A data frame with 2,116,800 rows and 8 variables:
\describe{
\item{year} {Year of the data}
\item{state} {State where the migration occurred}
\item{age} {Age to which belongs the data}
\item{CVE_GEO} {Geographical code of the state of occurrence}
\item{sex} {Sex of the migrants}
\item{emigrants} {Number of emigrants}
\item{immigrants} {Number of immigrants}
\item{type} {Type of migration: Interstate, International or Total}
}
}
\usage{
df_migration_expanded
}
\description{
A dataset containing the estimated migration by CONAPO between 1970 and 2050.
These estimations are disaggregated by year, state, sex and type of migration
(immigration or emigration).
}
\keyword{datasets}
