% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_mortrate_state_age_sex}
\alias{df_mortrate_state_age_sex}
\title{Population and deaths data of Mexico, from 1970 to 2050.}
\format{
A data frame with 882,420 rows and 7 variables:
\describe{
\item{year}{Year of the data}
\item{state}{State of the data}
\item{sex}{Sex of the data}
\item{CVE_GEO}{Geographical code of the state of the data}
\item{age}{Age of the data}
\item{population}{Estimated population by CONAPO}
\item{deaths}{Estimated deaths by CONAPO}
}
}
\usage{
df_mortrate_state_age_sex
}
\description{
A dataset containing the estimated population and deaths by CONAPO
between 1970 and 2050. These estimations are disaggregated by year, state
and sex
}
\keyword{datasets}
