% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aging_rate.R
\name{get_aging_rate}
\alias{get_aging_rate}
\title{Get aging rate}
\usage{
get_aging_rate(
  v_state = c("National", "Aguascalientes"),
  v_year = 2021,
  v_sex = c("Female", "Male", "Total"),
  v_age = c(0, 15, 24, 36),
  age_groups = FALSE
)
}
\arguments{
\item{v_state}{Character vector specifying the state(s) that the function
will return.}

\item{v_year}{Numeric vector that specifies the year(s) to return. Must have
numbers between 1985 and 2020. Default is 2021.}

\item{v_sex}{Character vector selecting sex. Options: Female, Male and Total.}

\item{age_groups}{Logical. Specifies whether to aggregate the output by
age groups.}
}
\value{
A demographic dataset containing the selected year
(group, when \code{year_groups = TRUE}), state, state code (CVE_GEO), sex,
the aging population and the aging rate.
}
\description{
\code{get_aging_rate} is a function that allows the user to get a demographic
dataset regarding the aging rate based on the given parameters.
}
