% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_births.R
\name{get_births}
\alias{get_births}
\title{Get births}
\usage{
get_births(v_states = "National", v_year = "2021", year_groups = TRUE)
}
\arguments{
\item{v_states}{State(s) of desired data.}

\item{v_year}{Year(s) of desired data. Must have numbers between 1970 and
2050. Default is 2021.}

\item{year_groups}{Specifies whether to aggregate the output by year groups.}

\item{export}{}
}
\value{
A demographic dataset based on the selected parameters.
}
\description{
\code{get_births} is a function that allows the user to
get a dataset of the number and rate of births based on different parameters.
}
\examples{
get_births(v_states = c("Aguascalientes", "Campeche"),
v_year = c(1995, 2021, 2046), year_groups = TRUE)

get_births(v_states = "National",  v_year = c(1985, 2005, 2046),
age_groups= FALSE)

}
