% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_births.R
\name{get_births_INEGI}
\alias{get_births_INEGI}
\title{Get births INEGI}
\usage{
get_births_INEGI(
  v_state = "National",
  v_year = 2021,
  v_sex = "Total",
  year_groups = TRUE
)
}
\arguments{
\item{v_state}{Character vector specifying the state(s) that the function
will return.}

\item{v_year}{Numeric vector that specifies the year(s) to return. Must have
numbers between 1985 and 2020. Default is 2021.}

\item{v_sex}{Vector selecting sex. Options: Female, Male and Total.}

\item{year_groups}{Logical. Specifies whether to aggregate the output by
year groups.}
}
\value{
A demographic dataset based on the selected parameters.
}
\description{
\code{get_births_INEGI} is a function that allows the user to get a
demographic dataset of the number and rate of births based on the given
parameters. The difference
}
\examples{
get_births(v_state = c("Aguascalientes", "Campeche"),
v_year = c(1995, 2021, 2046), year_groups = TRUE)

get_births(v_state = "National",  v_year = c(1985, 2005, 2046),
age_groups= FALSE)

}
