% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_death_rate.R
\name{get_death_rate}
\alias{get_death_rate}
\title{Get death rate}
\usage{
get_death_rate(
  v_state = "National",
  v_year = "2021",
  select_sex = unique(df_pop_state_age_sex$sex),
  v_init_age_grps,
  age_grps = T
)
}
\arguments{
\item{v_state}{State(s) of desired data}

\item{v_year}{Year(s) of desired data}

\item{select_sex}{Sex (Female/Male/Total)}

\item{v_init_age_grps}{Specifies the age bins to aggregate and return}

\item{age_grps}{(default is True)}

\item{export}{}
}
\value{
A dataset with the data selected
}
\description{
\code{get_death_rate} Get function that allows
user to get a dataset with the subset of different params
}
\examples{

}
