% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_migration.R
\name{get_migration}
\alias{get_migration}
\title{Load migration data}
\usage{
get_migration(
  v_state = "National",
  v_year = c(2000, 2010, 2025),
  v_sex = c("Female", "Male", "Total"),
  v_age = c(0, 15, 24, 36),
  v_type = c("Interstate", "International", "Total"),
  age_groups = TRUE
)
}
\arguments{
\item{v_state}{State(s) of desired data.}

\item{v_year}{Year(s) of desired data. Must have numbers between 1970 and
2050.}

\item{v_sex}{Vector selecting sex. Options: Female, Male and Total.}

\item{v_age}{Specifies the age bins to aggregate and return.}

\item{v_type}{Indicates whether type of migration to include. It can be
Interstate migration, International migration or Total migration.}

\item{age_groups}{Specifies whether to aggregate the output by age groups.}
}
\value{
A demographic dataset containing the selected year, the state, the
state code (CVE_GEO), the age (age group, when \code{age_groups = TRUE}), the
type of migration, the number of emigrants, the number of immigrants,
migration balance (immigrants - emigrants), and the rate of
emigration and immigration, respectively.
}
\description{
\code{get_migration} is a function that allows the user to get a demographic
dataset regarding migration based on the given parameters.
}
\examples{

get_migration(v_state = c("Yucatan", "Sonora"),
              v_year = 2015,
              v_sex = "Total",
              v_age = c(0, 25, 35, 45),
              v_type = "International",
              age_groups = TRUE)

get_migration(v_state = "Mexico City",
              v_year = c(2000, 2010),
              v_sex = "Female",
              v_age = c(0, 15, 35, 45, 75),
              v_type = c("Interstate", "International", "Total"),
              age_groups = FALSE)

}
