% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_births.R
\name{get_births}
\alias{get_births}
\title{Get births}
\usage{
get_births(v_state = "National", v_year = 2021, year_groups = TRUE)
}
\arguments{
\item{v_state}{Character vector specifying the state(s) that the function
will return.}

\item{v_year}{Numeric vector that specifies the year(s) to return. Must have
numbers between 1970 and 2050. Default is 2021.}

\item{year_groups}{Logical. Specifies whether to aggregate the output by
year groups.}
}
\value{
A demographic dataset containing the selected year
(group, when \code{year_groups = TRUE}), state, state code (CVE_GEO),
the number of births and the birth rate in each row,
}
\description{
\code{get_births} is a function that allows the user to get a demographic
dataset of the number and rate of births based on the given parameters.
}
\examples{
get_births(v_state = c("Aguascalientes", "Campeche"),
v_year = c(1995, 2021, 2046), year_groups = TRUE)

get_births(v_state = "National",  v_year = c(1985, 2005, 2046),
age_groups= FALSE)

}
