% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_births.R
\name{get_births_INEGI}
\alias{get_births_INEGI}
\title{Load births data, disaggregated by sex, from 1985 to 2020.}
\usage{
get_births_INEGI(
  v_state = "National",
  v_year = 2010,
  v_sex = c("Female", "Male", "Total"),
  year_groups = TRUE
)
}
\arguments{
\item{v_state}{Character vector specifying the state(s) that the function
will return.}

\item{v_year}{Numeric vector that specifies the year(s) to return. Must have
numbers between 1985 and 2020. Default is 2021.}

\item{v_sex}{Vector selecting sex. Options: Female, Male and Total.}

\item{year_groups}{Logical. Specifies whether to aggregate the output by
year groups.}
}
\value{
A demographic dataset containing the selected year
(group, when \code{year_groups = TRUE}), state, state code (CVE_GEO), sex,
the number of births the proportion of births (in relation to the total
births in that year), and the birth rate.
}
\description{
\code{get_births_INEGI} is a function that allows the user to get a
demographic dataset of the number and rate of births based on the given
parameters. This function returns the recorded births by INEGI and allows to
specify the sex.
}
\examples{


}
