% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_death_population.R
\name{get_death_population}
\alias{get_death_population}
\title{Get death population}
\usage{
get_death_population(
  v_state = "National",
  v_year = "2021",
  select_sex = unique(df_mortrate_state_age_sex$sex),
  v_init_age_grps,
  age_grps = T
)
}
\arguments{
\item{v_state}{State(s) of desired data}

\item{v_year}{Year(s) of desired data}

\item{select_sex}{Sex (Female/Male/Total)}

\item{v_init_age_grps}{Specifies the age bins to aggregate and return}

\item{age_grps}{(default is True)}

\item{export}{}
}
\value{
A dataset with the data select
}
\description{
\code{get_death_population} get function that allows user to
get a dataset with the subset of different params
}
\examples{
get_death_population(v_state = "Aguascalientes", v_year = "2021", select_sex = "Total",
v_init_age_grps = c("0","5","15","25","45", "55","65","70"), age_grps= T)
get_population(v_state = "Aguascalientes", v_year = c("2021", "2022"), select_sex = "Female",
v_init_age_grps = c("21", "22", "23"), age_grps= F)
}
