% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_migration.R
\name{get_migration}
\alias{get_migration}
\title{Get migration}
\usage{
get_migration(
  v_states = "Aguascalientes",
  v_year = c(2000, 2010, 2025),
  v_sex = c("Female", "Male", "Total"),
  v_age = c(0, 15, 24, 36),
  v_type = c("Interstate", "International", "Total"),
  age_groups = TRUE
)
}
\arguments{
\item{v_year}{Year(s) of desired data. Must have numbers between 1970 and
2050.}

\item{v_sex}{Vector selecting sex. Options: Female, Male and Total.}

\item{v_age}{Specifies the age bins to aggregate and return.}

\item{age_groups}{Specifies whether to aggregate the output by age groups.}

\item{v_state}{State(s) of desired data.}

\item{export}{}
}
\value{
A demographic dataset based on specified parameters.
}
\description{
\code{get_migration} Function that allows user to get a demographic dataset
regarding migration information based on the parameter set given by the
user
}
\examples{
get_population(v_states =  "Chiapas", v_year = 2015, v_sex = "Total",
v_age = c(0, 15, 45, 75), age_groups = FALSE)

get_population(v_states = c("Aguascalientes", "Campeche"),
v_year = c(2010, 2021), v_sex = "Male", v_age = c(0, 15, 45, 75),
age_groups = TRUE)

}
