% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population.R
\name{get_population}
\alias{get_population}
\title{Load population data disaggregated by year, state, sex and age.}
\usage{
get_population(
  v_state = "National",
  v_year = "2021",
  v_sex = "Total",
  v_age = c("0", "5", "15", "25", "45", "55", "65", "70"),
  age_groups = TRUE
)
}
\arguments{
\item{v_state}{State(s) of desired data.}

\item{v_year}{Year(s) of desired data. Must have numbers between 1970 and
2050.}

\item{v_sex}{Vector selecting sex. Options: Female, Male and Total.}

\item{v_age}{Specifies the age bins to aggregate and return.}

\item{age_groups}{Specifies whether to aggregate the output by age groups.}
}
\value{
A demographic dataset indicating the year, the state, the state code
(CVE_GEO), the age (group, when \code{age_groups = TRUE}), the population and
the proportion of each row, in relation to the selected year and sex.
}
\description{
\code{get_population} is a function that allows the user to get a demographic
dataset regarding population and population proportions based on the
given parameters.
}
\examples{
get_population(v_state =  "Chiapas", v_year = 2015, v_sex = "Total",
v_age = c(0, 15, 45, 75), age_groups = FALSE)

get_population(v_state = c("Aguascalientes", "Campeche"),
v_year = c(2010, 2021), v_sex = "Male", v_age = c(0, 15, 45, 75),
age_groups = TRUE)

}
